pyproject.toml
[project]
name = "predictive-maintenance"
version = "0.1.0"
description = "ML-based predictive maintenance for CNC machines"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "lightgbm>=4.1.0",
    "torch>=2.1.0",
    "tsfresh>=0.20.0",
    "pyarrow>=13.0.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "mlflow>=2.7.0",
    "optuna>=3.4.0",
    "joblib>=1.3.0",
    "tqdm>=4.66.0",
    "pydantic>=2.4.0",
    "evidently>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
